# Conflicts:
#	Backend/src/main/java/edu/teco/sensordatenbankmanagement/repository/DummyRepository.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/controllers/ObservationController.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/repository/Database.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/repository/ObservationRepository.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/repository/SensorRepository.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/services/ObservationService.java
#	Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/services/SensorService.java

+++ Backend/src/main/java/edu/teco/sensordatenbankmanagementsystem/models/Observation.java
@@ -13,10 +13,9 @@ import java.util.Map;
 @Data
 public class Observation {
 
-
     @Id
     @GeneratedValue
-    int id;
+    long id;
 
     private Map<String, String> properties;
